+++ CHANGELOG.md
@@ -68,6 +68,11 @@
   [SimplyDanny](https://github.com/SimplyDanny)
   [#6466](https://github.com/realm/SwiftLint/issues/6466)
 
+* Track identifiers declared in tuples to avoid false positives in the
+  `redundant_self` rule.  
+  [SimplyDanny](https://github.com/SimplyDanny)
+  [#6553](https://github.com/realm/SwiftLint/issues/6553)
+
 ## 0.63.2: High-Speed Extraction
 
 ### Breaking